Upload - Documentation Start
Documentation for assignment 1. Any updates made to Simulink block code or GUI code should be noted down mentioning what got changes, when did you change it (date and time) and why you changed it (explanation behind the motivation for the change). Change could be anything added, manipulated or removed. If possible reference Assignment 1 document and use direct quotes and references from the document for the motivation behind the changes. This will help us later on.